refactor: separate component directories by feature
