+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
 import { SessionProvider } from "next-auth/react";
-import MenuBar from "../components/MenuBar";
+import MenuBar from "../components/common/MenuBar";
 
 function MyApp({ Component, pageProps: { session, ...pageProps } }: AppProps) {
   return (
